Move handlers to wallet folder
