+++ app/app/utils.tsx
@@ -46,16 +46,6 @@ export function formatNumber(
   }).format(amount);
 }
 
-export async function fetchJson(url: string, options?: object) {
-  const res = await fetch(url, options);
-  if (!res.ok) {
-    throw new Error(
-      "fetchJson: http error: " + res.status + ": " + (await res.text()),
-    );
-  }
-  return await res.json();
-}
-
 export const addDollarSignAndSuffix = (value: number) => {
   if (value >= 1e6) {
     return `$${formatNumber(value / 1e6, 2, 2)}M`;
