+++ app/hooks/useWalletTokens.ts
@@ -1,51 +1,22 @@
-import { formatNumber, isERC20, normalizeAddress } from "@/app/utils";
+import { isERC20, normalizeAddress } from "@/app/utils";
 import {
   WalletTokensData,
   TokenData,
   ConnectedWalletsState,
 } from "@/utils/interfaces";
 import { ChainKey } from "@/utils/wallet/constants";
-import { getBalance, getPools, PoolDetail } from "@/midgard";
+import { getPools, PoolDetail } from "@/midgard";
 import { getChainKeyFromChain } from "@/utils/chain";
 import {
   assetFromString,
-  baseAmount,
-  baseToAsset,
 } from "@xchainjs/xchain-util";
-import { useUTXO } from "./useUTXO";
 import {
-  checkAndSwitchChain,
   initialWalletTokensData,
 } from "@/utils/wallet/balances";
 import { useQuery } from "@tanstack/react-query";
-import { SigningStargateClient } from "@cosmjs/stargate";
+import { chainHandlers } from "@/utils/wallet/handlers/handleBalance";
 
 export const useWalletTokens = (walletsState: ConnectedWalletsState) => {
-  // TODO: Avoid duplication of this condition between useUTXO and this line (https://linear.app/project-chaos/issue/YLD-141/consolidate-all-chain-configuration#comment-d10c7c6f)
-  const { getBalance: getBalanceBtc } = useUTXO({
-    chain: "BTC",
-    wallet: walletsState["Bitcoin"],
-  });
-  const { getBalance: getBalanceLtc } = useUTXO({
-    chain: "LTC",
-    wallet: walletsState["Litecoin"],
-  });
-  const { getBalance: getBalanceDoge } = useUTXO({
-    chain: "DOGE",
-    wallet: walletsState["Dogecoin"],
-  });
-  const { getBalance: getBalanceBch } = useUTXO({
-    chain: "BCH",
-    wallet: walletsState["BitcoinCash"],
-  });
-
-  const utxoBalancesHandlers = {
-    [ChainKey.BITCOIN]: getBalanceBtc,
-    [ChainKey.LITECOIN]: getBalanceLtc,
-    [ChainKey.DOGECOIN]: getBalanceDoge,
-    [ChainKey.BITCOINCASH]: getBalanceBch,
-  };
-
   const fetchWalletTokens = async () => {
     const updatedTokensData: WalletTokensData = {
       ...initialWalletTokensData,
@@ -134,55 +105,6 @@ export const useWalletTokens = (walletsState: ConnectedWalletsState) => {
     return { ...updatedTokensData };
   };
 
-  const getRuneBalance = async (walletAddress: string) => {
-    try {
-      const { data: runeBalance } = await getBalance({
-        path: {
-          address: walletAddress,
-        },
-      });
-      return runeBalance;
-    } catch {
-      return undefined;
-    }
-  };
-
-  const getUTXOInfo = async (
-    chainKey: "Bitcoin" | "Dogecoin" | "Litecoin" | "BitcoinCash",
-    walletAddress: string,
-  ) => {
-    // TODO: Remove once unify chains configurations (https://linear.app/project-chaos/issue/YLD-141/consolidate-all-chain-configuration#comment-d10c7c6f)
-    if (!walletsState) return;
-    if (!walletsState[chainKey]?.provider) return null;
-    const balance = await utxoBalancesHandlers[chainKey](walletAddress);
-    return {
-      balance: baseToAsset(balance.amount).amount().toNumber(),
-      formattedBalance: baseToAsset(balance.amount).amount().toString(),
-    };
-  };
-
-  const getCosmosBalance = async (walletAddress: string) => {
-    try {
-      const client = await SigningStargateClient.connect(
-        process.env.NEXT_PUBLIC_COSMOS_DIRECTORY_URL || "",
-      );
-      const atomDecimals = 6;
-      const balance = await client.getBalance(walletAddress, "uatom");
-      const baseAssetAmount = baseToAsset(
-        baseAmount(balance.amount, atomDecimals),
-      )
-        .amount()
-        .toNumber();
-      return {
-        balance: baseAssetAmount,
-        formattedBalance: formatNumber(baseAssetAmount, atomDecimals),
-      };
-    } catch (err) {
-      console.error("Error getting ATOM balance:", err);
-      return null;
-    }
-  };
-
   const getTokenBalances = async (walletTokensData: WalletTokensData): Promise<WalletTokensData> => {
     let newWalletTokensData: WalletTokensData = { ...walletTokensData };
   
@@ -206,61 +128,7 @@ export const useWalletTokens = (walletsState: ConnectedWalletsState) => {
       };
     };
   
-    const chainHandlers: Record<any, any> = {
-      [ChainKey.AVALANCHE]: async (address: string, provider: any, token: TokenData) => {
-        const info = await checkAndSwitchChain(
-          ChainKey.AVALANCHE,
-          address,
-          provider,
-          token.tokenAddress as `0x${string}`,
-        );
-        return { balance: baseToAsset(baseAmount(info?.balance.toString() || 0, 18)).amount().toNumber() };
-      },
-      [ChainKey.BSCCHAIN]: async (address: string, provider: any, token: TokenData) => {
-        const info = await checkAndSwitchChain(
-          ChainKey.BSCCHAIN,
-          address,
-          provider,
-          token.tokenAddress as `0x${string}`,
-        );
-        return { balance: baseToAsset(baseAmount(info?.balance.toString() || 0, 18)).amount().toNumber() };
-      },
-      [ChainKey.ETHEREUM]: async (address: string, provider: any, token: TokenData) => {
-        const info = await checkAndSwitchChain(
-          ChainKey.ETHEREUM,
-          address,
-          provider,
-          token.tokenAddress as `0x${string}`,
-        );
-        return { balance: baseToAsset(baseAmount(info?.balance.toString() || 0, 18)).amount().toNumber() };
-      },
-      [ChainKey.THORCHAIN]: async (address: string) => {
-        const info = await getRuneBalance(address);
-        const balance = info?.coins.find((coin) => coin.asset === "THOR.RUNE")?.amount || 0;
-        return { balance: Number(formatNumber(balance, 8)) };
-      },
-      [ChainKey.BITCOIN]: async (address: string) => {
-        const info = await getUTXOInfo(ChainKey.BITCOIN, address);
-        return { balance: info?.balance || 0 };
-      },
-      [ChainKey.LITECOIN]: async (address: string) => {
-        const info = await getUTXOInfo(ChainKey.LITECOIN, address);
-        return { balance: info?.balance || 0 };
-      },
-      [ChainKey.BITCOINCASH]: async (address: string) => {
-        const info = await getUTXOInfo(ChainKey.BITCOINCASH, address);
-        return { balance: info?.balance || 0 };
-      },
-      [ChainKey.DOGECOIN]: async (address: string) => {
-        const info = await getUTXOInfo(ChainKey.DOGECOIN, address);
-        return { balance: info?.balance || 0 };
-      },
-      [ChainKey.GAIACHAIN]: async (address: string) => {
-        const info = await getCosmosBalance(address);
-        return { balance: info?.balance || 0 };
-      },
-    };
-  
+    // TODO: We need to parallelize this requests (The EMV switch chain must be taken into account)
     for (const [chain, tokens] of Object.entries(walletTokensData)) {
       if (!walletsState || !walletsState[chain as ChainKey]) continue
     
@@ -304,7 +172,7 @@ export const useWalletTokens = (walletsState: ConnectedWalletsState) => {
   });
 
   return {
-    refreshBalances: refetch, // TODO: Avoid refresh all at once
+    refreshBalances: refetch,
     balanceList: walletBalances,
     isLoadingBalance: isFetching,
     isLoadingTokenList: isFetchingWalletTokens,
