+++ app/utils/wallet/handlers/handleConnect.ts
