+++ app/utils/wallet/handlers/handleTransfer.ts
@@ -1,5 +1,5 @@
 import { Asset, baseAmount, baseToAsset } from "@xchainjs/xchain-util";
-import { WalletState } from "../interfaces";
+import { WalletState } from "../../interfaces";
 import { Client as BitcoinClient } from "@xchainjs/xchain-bitcoin";
 import * as Bitcoin from 'bitcoinjs-lib'
 
@@ -54,36 +54,33 @@ export const transferUTXO = async (
         const fromHexString = (hexString: any) =>
           Uint8Array.from(hexString.match(/.{1,2}/g).map((byte: any) => parseInt(byte, 16)));
         const psbtUnsigned = Bitcoin.Psbt.fromBase64(rawUnsignedTx)
-        try {
-          const signedPSBTBytes: Uint8Array = await wallet.provider.signPSBT(
-            fromHexString(psbtUnsigned.toHex()),
-            {
-              inputsToSign: [{
-                address: transferParams.from,
-                signingIndexes: inputs.map(input => input.index),
-              }]
-            }
-          );
+        const signedPSBTBytes: Uint8Array = await wallet.provider.signPSBT(
+          fromHexString(psbtUnsigned.toHex()),
+          {
+            inputsToSign: [{
+              address: transferParams.from,
+              signingIndexes: inputs.map(input => input.index),
+            }]
+          }
+        );
 
-          const psbtSigned = Bitcoin.Psbt.fromBuffer(signedPSBTBytes)
-          psbtSigned.finalizeAllInputs()
+        const psbtSigned = Bitcoin.Psbt.fromBuffer(signedPSBTBytes)
+        psbtSigned.finalizeAllInputs()
 
-          const txHex = psbtSigned.extractTransaction().toHex()
+        const txHex = psbtSigned.extractTransaction().toHex()
 
-          psbtSigned.extractTransaction().getId()
+        psbtSigned.extractTransaction().getId()
 
-          const hash = clientBuilder.broadcastTx(txHex)
-          return hash
-        } catch(e) {
-          console.error('Error transfer UTXO using Phantom', e)
-          throw Error('Error transfer UTXO using Phantom')
-        }
+        const hash = clientBuilder.broadcastTx(txHex)
+        return hash
       case "okx":
-        const { txHash } = wallet.provider.send({ 
+        const value = baseToAsset(baseAmount(transferParams.amount.amount, transferParams.amount.decimals)).amount().toString()
+        const { txHash } = await wallet.provider.send({ 
           from: transferParams.from, 
           to: transferParams.recipient, 
-          value: baseToAsset(baseAmount(transferParams.amount.amount, transferParams.amount.decimals)).amount().toString(), 
+          value: value, 
           memo: transferParams.memo,
+          memoPos: 0,
           satBytes: transferParams.feeRate
         })
         return txHash
